Rework API for reading input
It seems more intuitive to have a consistent return
value for issues with html_recv and html_form_read.
This reworked indicating a close request to a stateful
function called html_close_requested, similar to feof
in stdio.h.  Fix #44.

Also, it was already a breaking change to html_read_form,
so it was renamed to html_form_read to fix #43.