Spin off thread to clean up old sessions
It's at least common in dev to kill the server with Ctrl+C.
This prevents the cleanup code from being run that would
delete session directories of previously uploaded content.
Now, when the server starts up, it'll spin off a thread to
scan directories, attempt to lock them, and delete them.

Fix #20