Prevent default for snake keys
This is necessary because WkWebView hosting this
web page will beep on every arrow key if the default
action isn't prevented. Seems like this probably happens
for free when an interactive element like an <input /> is
focused by the browser, but generic window-level event
listeners likely need to implement it themselves.