Rework package tests
This split the upload/creating of a package into a separate
workflow call for better organization. Then, this combined
the testing of the two different packages so that we could
make sure that the client is installed before running the
server test, since the server library depends on the client.