Rework window_id to be session_id
This is a significant change. The reason this is done is
so that a window can be associated with a server session.
That is important so that something like the host
application can know exactly which session a window operation
ends with. Originally I thought it would be neat to have
a 1 to many relationship of session ID to window ID, where
a session could open many windows. That seems like it breaks
the simplistic model of the html_forms protocol. If such
generic applications are needed, we should probably make
more general purpose and robust protocols for web server
and browser automation that are out of scope of the simple
html_forms protocol.

From another angle, html_forms is supposed to make web app
development somewhat on par with printf style programs from
ease of development.  Printf programs don't have multiple
consoles to write to unless they're super weird. In a similar
sense, if a single window doesn't suffice for an html_forms
app, then the app probably should take on a more complex
protocol.